Updated Vs2008 and Vs2010 settings. Added some missing files to VS2008, player2 compiles now. Disabled warning 4996 (fopen, sprintf etc. is deprecated use _s) and other small changes to the project files. Disabled Win32 Entrypoint because this breaks compilation under VS. Added Debug and Release output folders to ignore list.

+++ src/game_map.cpp
@@ -265,7 +265,7 @@ void Game_Map::StartFogOpacityChange(int opacity, int duration) {
 	fog_opacity_target = opacity * 1.0;
 	fog_opacity_duration = duration;
 	if (fog_opacity_duration == 0) {
-		fog_opacity = fog_opacity_target;
+		fog_opacity = (int)fog_opacity_target;
 	}
 }
